Fix tests (#25)
The main problem that this fixes is the unability to run `make test`.
Before this fix, it'd fail with this error:

```
can't load package: package github.com/andrecronje/lachesis: build constraints exclude all Go files in /opt/go/src/github.com/andrecronje/lachesis
```

The problem was that in the context of the makefile, `glide novendor`
would return the following output:

```
./cmd/...
./src/...
./test/...
.
```

When you try to pass the fourth line, `.`, to `xargs`, that would upset
go, it's not the kind of directory it would accept. As a quick patch, I
removed from the ouput of `glide novendor` all the references to `.`.

I also noticed some misconfigurations. The main one is that the lack of
version in the `lfshook` package would prevent it from being added into
the vendor folder, causing a bunch of weird to errors: Because it wasn't
in vendor, it would `go` would get the version in github and consider
the types there different from the types in the packages that depend
from `lfshook` but are in `vendor`.

I also fixed a bunch of tests that were happening. I didn't fix all of
them, as the majority seemed to be caused by in progress test for which
I don't have enough context nor knowledge (yet) to be able to fix them.