Refactor reaction to EVM callbacks
Using EVM callbacks CaptureEnter and CaptureExit, which were
not available when this code was written. Because of that stop
using CaptureState as for transaction tracing is not needed to
capture every single instruction.
That leads to not using EVM stack and memory to get input data
for inner calls and results for return, stop and revert. This data
is now available directly in the callback arguments.